feat: refactor TransactionForm and add TransactionHeader component
- Updated imports in TransactionForm to use local paths for TransactionHeader, CategoryDropdown, and TransactionNotes.
- Introduced a new TransactionHeader component to display transaction details including date, account name, and amount.
- Modified TransactionNotes to accept a nullable string for the onChange prop, enhancing type safety.
